Rename: 기존 icon의 png파일을 svg로 변경 화면 확대 시 이미지 깨짐을 완화
