+++ src/components/button/button.style.jsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import socialImg from '../../assets/image/social_login_sprites.png';
-import chat from '../../assets/icon/message-circle.png';
-import share from '../../assets/icon/share.png';
+import chat from '../../assets/icon/message-circle.svg';
+import share from '../../assets/icon/share.svg';
 
 export const SocialLoginButton = styled.button`
 	display: block;
